rename Notification.js-> notification.js
