+++ src/background/notification.js
